@@ -1,19 +1,19 @@
-// swift-tools-version: 5.8
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let package = Package(
-    name: "swift-redis",
+    name: "SwiftRedis",
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
         // Targets can depend on other targets in this package and products from dependencies.
         .executableTarget(
-            name: "swift-redis",
+            name: "SwiftRedis",
             path: "Sources"),
         .testTarget(
             name: "swift-redis-tests",
-            dependencies: ["swift-redis"],
+            dependencies: ["SwiftRedis"],
             path: "Tests"),
     ]
 )
